+++ Plugins/DryreLUIEssentialsPlugin/Source/DryreLUIEssentialsPlugin/Public/ADLManager.h
@@ -62,7 +62,7 @@ int adlGetGPUClockSpeedADL();
 int adlGetGPUVRAMClockSpeedADL();
 
 // Display the system time stamp (in ms)
-int64 adlGetTimeStampADL();
+int adlGetTimeStampADL();
 
 // Display GPU temperature(in °C)
 int adlGetGPUTemperatureADL();
